Move prototype-specific utilities to subpackage
