+++ gist_memory/__init__.py
@@ -57,16 +57,16 @@ _lazy_map = {
     "AgenticChunker": "gist_memory.chunker",
     "DEFAULT_BRAIN_PATH": "gist_memory.config",
     "LocalChatModel": "gist_memory.local_llm",
-    "render_five_w_template": "gist_memory.canonical",
-    "MemoryCueRenderer": "gist_memory.memory_cues",
-    "ConflictFlagger": "gist_memory.conflict_flagging",
-    "ConflictLogger": "gist_memory.conflict_flagging",
-    "SimpleConflictLogger": "gist_memory.conflict",
+    "render_five_w_template": "gist_memory.prototype.canonical",
+    "MemoryCueRenderer": "gist_memory.prototype.memory_cues",
+    "ConflictFlagger": "gist_memory.prototype.conflict_flagging",
+    "ConflictLogger": "gist_memory.prototype.conflict_flagging",
+    "SimpleConflictLogger": "gist_memory.prototype.conflict",
     "ExperimentConfig": "gist_memory.experiment_runner",
     "run_experiment": "gist_memory.experiment_runner",
     "HistoryExperimentConfig": "gist_memory.history_experiment",
     "run_history_experiment": "gist_memory.history_experiment",
-    "negation_conflict": "gist_memory.conflict",
+    "negation_conflict": "gist_memory.prototype.conflict",
     "TalkSessionManager": "gist_memory.talk_session",
     "load_agent": "gist_memory.utils",
     "tokenize_text": "gist_memory.token_utils",
